fix(ci): clear mypy + clippy failures latent after #53 merge
CI on main (run 25642714901) failed `Lint + typecheck` (mypy) and `Rust`
(`cargo clippy -- -D warnings`) once the ruff failure was unblocked.
These were latent issues from fc268a1.

mypy (6 errors):
- nn/presolve.py: predeclare used_lb/used_ub as Optional[np.ndarray]
  so the if/else branches join to a union instead of being inferred
  as ndarray from the first branch.
- _jax/presolve_pipeline.py: annotate agg_total as `dict` (matching
  rc_total/rcf_total/scaling_total) — the mixed int/list literal was
  being inferred as dict[str, object].

clippy (8 errors, -D warnings):
- presolve/fbbt_fp.rs: replace needless range loop with
  in_queue.iter_mut().enumerate().
- presolve/redundancy.rs: same idiom, iter().enumerate().skip(i+1).
- presolve/symmetry.rs: drop six redundant `*deg as u32` casts —
  deg is already u32.

All 155 presolve unit tests still pass; mypy and clippy now clean
locally.

+++ crates/discopt-core/src/presolve/symmetry.rs
@@ -145,7 +145,7 @@ pub fn detect_symmetries(model: &ModelRepr) -> (Vec<Orbit>, SymmetryStats) {
                 let mut tokens: Vec<(u8, u32)> = Vec::new();
                 for (fid, deg) in &m.factors {
                     if *fid == leaf {
-                        tokens.push((0xFE, *deg as u32));
+                        tokens.push((0xFE, *deg));
                     } else if let ExprNode::Variable {
                         index: other_idx,
                         size: 1,
@@ -159,12 +159,12 @@ pub fn detect_symmetries(model: &ModelRepr) -> (Vec<Orbit>, SymmetryStats) {
                                 VarType::Binary => 2,
                                 VarType::Integer => 3,
                             };
-                            tokens.push((ot, *deg as u32));
+                            tokens.push((ot, *deg));
                         } else {
-                            tokens.push((0xFD, *deg as u32));
+                            tokens.push((0xFD, *deg));
                         }
                     } else {
-                        tokens.push((0xFD, *deg as u32));
+                        tokens.push((0xFD, *deg));
                     }
                 }
                 tokens.sort();
@@ -195,9 +195,9 @@ pub fn detect_symmetries(model: &ModelRepr) -> (Vec<Orbit>, SymmetryStats) {
                 let mut tokens: Vec<(u8, u32)> = Vec::new();
                 for (fid, deg) in &m.factors {
                     if *fid == leaf {
-                        tokens.push((0xFE, *deg as u32));
+                        tokens.push((0xFE, *deg));
                     } else {
-                        tokens.push((0xFD, *deg as u32));
+                        tokens.push((0xFD, *deg));
                     }
                 }
                 tokens.sort();
